+++ src/__screenshot_tests__/text-field-screenshot-test.tsx
@@ -5,51 +5,51 @@ const TESTABLE_DEVICES: Array<Device> = ['MOBILE_IOS', 'DESKTOP'];
 
 test.each(TESTABLE_DEVICES)('Default textfield appears properly on %s', async (device) => {
     const page = await openStoryPage({
-        section: 'Components|Forms/TextField',
+        section: 'Components|Forms/TextField (deprecated)',
         name: 'Variants',
         device,
     });
 
-    const image = await page.screenshot();
+    const image = await page.screenshot({fullPage: true});
     expect(image).toMatchImageSnapshot();
 });
 
 test.each(TESTABLE_DEVICES)('Default textfield appears properly (focus) on %s', async (device) => {
     const page = await openStoryPage({
-        section: 'Components|Forms/TextField',
+        section: 'Components|Forms/TextField (deprecated)',
         name: 'Variants',
         device,
     });
 
-    page.click(await screen.findByLabelText('Normal field (opcional)'));
+    page.click(await screen.findByLabelText('Normal field'));
 
-    const image = await page.screenshot();
+    const image = await page.screenshot({fullPage: true});
     expect(image).toMatchImageSnapshot();
 });
 
 test.each(TESTABLE_DEVICES)('Default textfield appears properly (typing) on %s', async (device) => {
     const page = await openStoryPage({
-        section: 'Components|Forms/TextField',
+        section: 'Components|Forms/TextField (deprecated)',
         name: 'Variants',
         device,
     });
 
-    await page.type(await screen.findByLabelText('Normal field (opcional)'), 'hello moto', {delay: 100});
+    await page.type(await screen.findByLabelText('Normal field'), 'hello moto', {delay: 100});
 
-    const image = await page.screenshot();
+    const image = await page.screenshot({fullPage: true});
     expect(image).toMatchImageSnapshot();
 });
 
 test.each(TESTABLE_DEVICES)('Default textfield appears properly (typing and blur) on %s', async (device) => {
     const page = await openStoryPage({
-        section: 'Components|Forms/TextField',
+        section: 'Components|Forms/TextField (deprecated)',
         name: 'Variants',
         device,
     });
 
-    await page.type(await screen.findByLabelText('Normal field (opcional)'), 'hello moto', {delay: 100});
-    await page.click((await screen.findAllByLabelText('Multiline (opcional)'))[0]);
+    await page.type(await screen.findByLabelText('Normal field'), 'hello moto', {delay: 100});
+    await page.click((await screen.findAllByLabelText('Multiline'))[0]);
 
-    const image = await page.screenshot();
+    const image = await page.screenshot({fullPage: true});
     expect(image).toMatchImageSnapshot();
 });