+++ src/form-date-field.tsx
@@ -4,38 +4,19 @@ import {TextFieldBase} from './text-field-base';
 
 import type {CommonFormFieldProps} from './form';
 
-export interface SimpleFormTextFieldProps extends CommonFormFieldProps {
-    type?: 'text'; // @deprecated, this will be the only allowed type for FormTextFields
+export interface FormDateFieldProps extends CommonFormFieldProps {
     onChangeValue?: (value: string, rawValue: string) => void;
-    multiline?: boolean;
-    prefix?: React.ReactNode;
-    endIcon?: React.ReactNode;
-    getSuggestions?: (value: string) => Array<string>;
 }
 
-/**
- * @deprecated
- */
-export interface OtherFormTextFieldProps extends CommonFormFieldProps {
-    type?: 'password' | 'integer' | 'decimal' | 'date';
-    onChangeValue?: (value: string, rawValue: string) => void;
-    multiline?: undefined;
-    prefix?: React.ReactNode;
-    endIcon?: React.ReactNode;
-}
-
-export type FormTextFieldProps = SimpleFormTextFieldProps | OtherFormTextFieldProps;
-
-export const FormTextField: React.FC<FormTextFieldProps> = ({
+export const FormDateField: React.FC<FormDateFieldProps> = ({
     disabled,
     error,
     helperText,
     name,
     optional,
-    type = 'text',
     validate,
-    onChangeValue,
     onChange,
+    onChangeValue,
     onBlur,
     value,
     ...rest
@@ -51,38 +32,28 @@ export const FormTextField: React.FC<FormTextFieldProps> = ({
         register,
     } = useForm();
 
-    // TODO Remove: APPS-XXXX
-    React.useEffect(() => {
-        if (process.env.NODE_ENV !== 'production') {
-            if (type !== 'text') {
-                console.error(
-                    'FormTextFields with a type different than "text" are deprecated. Please use another FormField component.' +
-                        '\nSee: https://mistica-web.now.sh/?path=/story/components-forms-formfields--types-uncontrolled'
-                );
-            }
-        }
-    }, [type]);
+    const processValue = (value: string) => value;
 
     return (
         <TextFieldBase
             {...rest}
+            shrinkLabel
+            type="date"
             inputRef={(field) => register({name, field, validate})}
             disabled={disabled || formStatus === 'sending'}
             error={error || !!formErrors[name]}
             helperText={formErrors[name] || helperText}
-            type={type}
             name={name}
             required={!optional}
             value={value ?? rawValues[name] ?? (rest.defaultValue !== undefined ? undefined : '')}
             onChange={(event) => {
                 const rawValue = event.currentTarget.value;
-                const value = rawValue;
+                const value = processValue(rawValue);
                 setRawValue({name, value: rawValue});
                 setValue({name, value});
-
+                setFormError({name, error: ''});
                 onChange?.(event);
                 onChangeValue?.(value, rawValue);
-                setFormError({name, error: ''});
             }}
             onBlur={(e) => {
                 setFormError({name, error: validate?.(values[name], rawValues[name])});