+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # CHANGELOG
 
+## Release pending
+
+- Form fields refactor:
+  - Deprecate `TextField` (will be removed in a future release)
+  - Deprecate `FormTextField` with types different than `"text"` (those usages will be removed in a future
+    release)
+  - New semantic FormField components added:
+    - `FormDateField`
+    - `FormDecimalField`
+    - `FormEmailField`
+    - `FormIntegerField`
+    - `FormPasswordField`
+  - All FormField components can now be used inside a `<Form>` and standalone (connected or unconnected)
+  - `name` prop is in FormFields is now mandatory
+
 ## v.4.1.0 - 2020-07-21
 
 - Removed border from `BoxedRowList` component, moved it to `BoxedRow` component